feat: added functions for creating player, club and leagues

+++ src/include/LiFM-Data/League.h
@@ -8,7 +8,7 @@
 
 struct League
 {
-  int leagueID;
+  size_t leagueID;
   std::vector<int> clubs;
 };
